refactor: moved directory domain and its files
