+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/Queue.java
@@ -1,6 +1,6 @@
-package queue;
+package com.StationManager.app.domain;
 
-import client.Client;
+import com.StationManager.app.domain.client.Client;
 
 import java.util.ArrayList;
