+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/Transaction.java
@@ -1,4 +1,4 @@
-package transaction;
+package com.StationManager.app.domain;
 
 import client.Client;
