+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/client/Client.java
@@ -1,4 +1,4 @@
-package client;
+package com.StationManager.app.domain.client;
 
 import java.awt.*;
