+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/trainstation/Direction.java
@@ -1,0 +1,8 @@
+package com.StationManager.app.domain.trainstation;
+
+public enum Direction {
+    Up,
+    Down,
+    Left,
+    Right
+}
