+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/trainstation/Position.java
@@ -1,4 +1,4 @@
-package trainstation;
+package com.StationManager.app.domain.trainstation;
 
 import java.awt.*;
