+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/trainstation/TrainStation.java
@@ -1,4 +1,4 @@
-package trainstation;
+package com.StationManager.app.domain.trainstation;
 
 public class TrainStation {
     private Integer id;
